Add support for multi-dimension critical css.
TLDR

This is useful when you want to deliver multiple dimensions
of critical css. Take for instance when your site has different elements
visible at different resolutions. Prior to this work you were limited to
multiple critical passes and inclusions or simply picking one
resolution.

Why is this better?

I am glad you asked yourself that. Since we use
clean-css we can intellegently merge the multiple resolution snapshots
making for smaller and better multi snapshot rule generation.

Why did I do this?

While working on a website I noticed that we were getting a flash of
styles after the normal stylesheet was loaded. This was not so bad on a
computer screen (since that was the dimension that we targeted) however
when on a phone it was quite a jarring experience.